requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
colorama==0.4.6
distro==1.8.0
Django==4.2.7
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
openai==1.3.2
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
python-decouple==3.7
pytz==2023.3.post1
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3